+++ changelog.md
@@ -14,3 +14,4 @@ vX.X.X
 - Added layout options for RX/TX tab, closes #81.
 - Changed the "UTF-8" decoding option to "ASCII", closes #83.
 - Added buffer limit for TX and RX data, closes #86.
+- Added auto-scroll to status message pane, closes #87.